# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API Agenda de Contacts",
        "description": "Documentação da API de Agenda de Contacts",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://127.0.0.1:8000",
            "description": "Servidor Local"
        }
    ],
    "paths": {
...
        "/api/contacts": {
            "get": {
                "tags": [
                    "Contacts"
                ],
                "summary": "Lista todos os Contacts",
                "operationId": "bc39494d59500e2d75e6f6648b43c8de",
                "parameters": [
                    {
                        "name": "search",
                        "in": "query",
                        "description": "Termo de pesquisa para filtrar Contacts pelo name",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "phone",
                        "in": "query",
                        "description": "Buscar pelo phone",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "favorite",
                        "in": "query",
                        "description": "Filtrar favorites",
                        "required": false,
                        "schema": {
                            "type": "boolean"
                        }
                    },
                    {
                        "name": "sort",
                        "in": "query",
                        "description": "Ordenar por name ou created_at",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "direction",
                        "in": "query",
                        "description": "asc ou desc",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "per_page",
                        "in": "query",
                        "description": "Quantidade de registros por página",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Lista de Contacts retornada com sucesso"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Contacts"
                ],
                "summary": "Cria um novo Contact",
                "operationId": "779a1f2268bae1138e7ea7843dd88d90",
                "parameters": [
                    {
                        "name": "name",
                        "in": "query",
                        "description": "name do Contact",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "phone",
                        "in": "query",
                        "description": "phone do Contact",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "email",
                        "in": "query",
                        "description": "Email do Contact",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "favorite",
                        "in": "query",
                        "description": "Indica se o Contact é favorite",
                        "required": false,
                        "schema": {
                            "type": "boolean"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Contact criado com sucesso"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/contacts/{Contact}": {
            "get": {
                "tags": [
                    "Contacts"
                ],
                "summary": "Exibe um Contact",
                "operationId": "d0d60696adc4804f953f5b3e37407048",
                "parameters": [
                    {
                        "name": "Contact",
                        "in": "path",
                        "description": "ID do Contact",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Contact retornado com sucesso"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Contacts"
                ],
                "summary": "Atualiza um Contact",
                "operationId": "3c4aae286067269fe8c8b3162c0059d1",
                "parameters": [
                    {
                        "name": "Contact",
                        "in": "path",
                        "description": "ID do Contact",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "name",
                        "in": "query",
                        "description": "name do Contact",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "phone",
                        "in": "query",
                        "description": "phone do Contact",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "email",
                        "in": "query",
                        "description": "Email do Contact",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "favorite",
                        "in": "query",
                        "description": "Indica se o Contact é favorite",
                        "required": false,
                        "schema": {
                            "type": "boolean"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Contact atualizado com sucesso"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Contacts"
                ],
                "summary": "Remove um Contact",
                "description": "Remove the specified resource from storage.",
                "operationId": "96ed7c648f93c6e4a2f97abb672b9abc",
                "parameters": [
                    {
                        "name": "Contact",
                        "in": "path",
                        "description": "ID do Contact",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Contact removido com sucesso"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/contacts/{Contact}/favorite": {
            "patch": {
                "tags": [
                    "Contacts"
                ],
                "summary": "Atualiza o status de favorite de um Contact",
                "operationId": "c7995e8033cef9111bb944cc47af9e4e",
                "parameters": [
                    {
                        "name": "Contact",
                        "in": "path",
                        "description": "ID do Contact",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "favorite",
                        "in": "query",
                        "description": "Indica se o Contact é favorite",
                        "required": true,
                        "schema": {
                            "type": "boolean"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Status de favorite atualizado com sucesso"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "sanctum": {
                "type": "apiKey",
                "description": "Autenticação via Laravel Sanctum. O frontend usa sessão de SPA (cookie + CSRF); integrações externas podem usar um token pessoal no formato: Bearer <token>.",
                "name": "Authorization",
                "in": "header"
            }
        }
    },
...
}